Extract RokuDeployOptions into its own function.
