+++ src/RokuDeploy.ts
@@ -10,6 +10,7 @@ import { promisify } from 'util';
 const globAsync = promisify(glob);
 
 import { util } from './util';
+import { RokuDeployOptions, FileEntry } from './RokuDeployOptions';
 
 export class RokuDeploy {
     //store the import on the class to make testing easier
@@ -884,129 +885,6 @@ export class RokuDeploy {
             archive.finalize();
         });
     }
-}
-
-export interface RokuDeployOptions {
-    /**
-     * A full path to the folder where the zip/pkg package should be placed
-     * @default './out'
-     */
-    outDir?: string;
-
-    /**
-     * The base filename the zip/pkg file should be given (excluding the extension)
-     * @default 'roku-deploy'
-     */
-    outFile?: string;
-
-    /**
-     * The root path to the folder holding your Roku project's source files (manifest, components/, source/ should be directly under this folder)
-     * @default './'
-     */
-    rootDir?: string;
-
-    /**
-     * An array of source file paths, source file globs, or {src,dest} objects indicating
-     * where the source files are and where they should be placed
-     * in the output directory
-     * @default [
-            'source/**\/*.*',
-            'components/**\/*.*',
-            'images/**\/*.*',
-            'manifest'
-        ],
-     */
-    files?: FileEntry[];
-
-    /**
-     * Set this to true to prevent the staging folder from being deleted after creating the package
-     * @default false
-     */
-    retainStagingFolder?: boolean;
-
-    /**
-     * Should the zipped package be retained after deploying to a roku. If false, this will delete the zip after a deployment.
-     * @default true
-     */
-    retainDeploymentArchive?: boolean;
-
-    /**
-     * The path where roku-deploy should stage all of the files right before being zipped. defaults to ${outDir}/.roku-deploy-staging
-     */
-    stagingFolderPath?: string;
-
-    /**
-     * The IP address or hostname of the target Roku device.
-     * @required
-     * @example '192.168.1.21'
-     *
-     */
-    host?: string;
-
-    /**
-     * The port that should be used when installing the package. Defaults to 80.
-     * This is mainly useful for things like emulators that use alternate ports,
-     * or when publishing through some type of port forwarding configuration.
-     */
-    packagePort?: number;
-
-    /**
-     * When publishing a side loaded channel this flag can be used to enable the socket based BrightScript debug protocol. Defaults to false.
-     * More information on the BrightScript debug protocol can be found here: https://developer.roku.com/en-ca/docs/developer-program/debugging/socket-based-debugger.md
-     */
-    remoteDebug?: boolean;
-
-    /**
-     * The port used to send remote control commands (like home press, back, etc.). Defaults to 8060.
-     * This is mainly useful for things like emulators that use alternate ports,
-     * or when sending commands through some type of port forwarding.
-     */
-    remotePort?: number;
-
-    /**
-     * The username for the roku box. This will always be 'rokudev', but allows to be overridden
-     * just in case roku adds support for custom usernames in the future
-     * @default 'rokudev'
-     */
-    username?: string;
-
-    /**
-     * The password for logging in to the developer portal on the target Roku device
-     * @required
-     */
-    password?: string;
-
-    /**
-     * The password used for creating signed packages
-     * @required
-     */
-    signingPassword?: string;
-
-    /**
-     * Path to a copy of the signed package you want to use for rekeying
-     * @required
-     */
-    rekeySignedPackage?: string;
-
-    /**
-     * Dev ID we are expecting the device to have. If supplied we check that the dev ID returned after keying matches what we expected
-     */
-    devId?: string;
-
-    /**
-     * If true we increment the build number to be a timestamp in the format yymmddHHMM
-     */
-    incrementBuildNumber?: boolean;
-
-    /**
-     * If true we convert to squashfs before creating the pkg file
-     */
-    convertToSquashfs?: boolean;
-
-    /**
-     * If true, the publish will fail on compile error
-     */
-    failOnCompileError?: boolean;
 }
 
 export interface ManifestData {
